Create Coins
Alice and Bob are playing a game using a bunch of coins.

The players pick several coins out of the bunch in turn. Each time a player is allowed to pick 1, 2 or 4 coins, and the player that gets the last coin is the winner. Assume that both players are very smart and he/she will try his/her best to work out a strategy to win the game.

For example, if there are 2 coins and Alice is the first player to pick, she will definitely pick 2 coins and win.

If there are 3 coins and Alice is still the first player to pick, no matter she picks 1 or 2 coins, Bob will get the last coin and win the game.

Given the number of coins and assuming Alice starts every time, you are required to write a program to calculate whether she will win the game or lose.

Input Format Input a single integer n, the number of coins. (1 <= n <= 100)

Output Format Output the name of the player who wins the game ("Alice" or "Bob") (first letter should be capital and print without quotes " ")

Example:

Input: 5

Output: Alice

Input : 30

Output: Bob